日付の変換処理をExtensionに移行
